requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
Pillow>=10.0.0
pytesseract>=0.3.10